+++ drivers/usb/gadget/function/f_mass_storage.c
@@ -849,7 +849,7 @@ static int do_read_cd(struct fsg_common *common)
 	loff_t			file_offset, file_offset_tmp;
 	unsigned int		amount;
 	unsigned int		partial_page;
-	ssize_t			nread;
+	unsigned int		nread;
 
 	u32 nb_sectors, transfer_request;
