+++ include/linux/fs.h
@@ -423,7 +423,7 @@ struct address_space {
 #if defined(CONFIG_MMC_DW_FMP_ECRYPT_FS) || defined(CONFIG_UFS_FMP_ECRYPT_FS)
 	unsigned char		*iv;		/* iv */
 	unsigned char		*key;		/* key */
-	unsigned long		key_length;	/* key length */
+	unsigned int		key_length;	/* key length */
 	char			*alg;		/* algorithm */
 	pgoff_t			sensitive_data_index;	/* data starts here */
 	struct crypto_hash	*hash_tfm;	/* hash transform */
